fix PyInt formatting with long long
pypy/rpython/translator/c/test/test_standalone.py mostly passes despite HEAP CORRUPTION DETECTED error boxes, except:
    TestStandalone::
        test_inhibit_tail_call
        test_set_length_fraction
        test_stack_criticalcode
        test_int_manipulation
    TestThread::
        test_stack_size
        test_thread_and_gc
    TestShared::
        test_entrypoint (crashes)

+++ Objects/longobject.c
@@ -348,7 +348,7 @@ int
 _PyLong_AsInt(PyObject *obj)
 {
     int overflow;
-    long result = PyLong_AsLongAndOverflow(obj, &overflow);
+    long long result = PyLong_AsLongAndOverflow(obj, &overflow);
     if (overflow || result > INT_MAX || result < INT_MIN) {
         /* XXX: could be cute and give a different
            message for overflow == -1 */
